+++ CHANGES.rst
@@ -1,6 +1,13 @@
 CHANGES
 =======
 
+master (unreleased)
+-------------------
+
+* Fix use of a secret with a percent symbol in it. Thanks Joey Wilhelm for the
+  report. Fixes GH-1.
+
+
 0.1 (2015.05.27)
 ----------------